refactor: centralize ACCESS_TOKEN_KEY in constants for consistency
