+++ src/app/api/oauth/kakao/token/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server"
 
-import { ACCESS_TOKEN_KEY } from "../route"
+import { ACCESS_TOKEN_KEY } from "@/constants/cookies"
 
 export async function GET(req: NextRequest) {
   const tokenCookie = req.cookies.get(ACCESS_TOKEN_KEY)
