+++ src/constants/cookies.ts
@@ -1,0 +1,1 @@
+export const ACCESS_TOKEN_KEY = "access_token"
