--- src/utils/storage/index.ts
@@ -1,7 +1,0 @@
-import { LocalStorageManager } from "./local-storage"
-
-export const ACCESS_TOKEN_KEY = "@@access-token"
-export const accessTokenStorage = new LocalStorageManager<string>(
-  ACCESS_TOKEN_KEY,
-  null,
-)
